* device/include/malloc.h: removed redundant __reentrant prototypes * device/lib/_mullong.c: added working xstack variant in asm (C version   doesn't pass regression tests) * device/lib/bpx.c: used __data and made bpx char for mcs51 * src/SDCCast.c (decorateType): removed unused GPTYPE_IDATA,   (createFunction): fixed bug with xstackPtr * src/SDCCcse.c: corrected comments * src/SDCCopt.c (convertToFcall): fixed warning in MSVC,   (killDeadCode, eBBlockFromiCode): removed unused code * src/SDCCsymt.h: removed unused GPTYPE_GPTR and GPTYPE_IDATA,   corrected comments * src/mcs51/gen.c (aopForSym, aopPut, toBoolean, unsaveRegisters,   assignResultValue, genCall, genFunction, genEndFunction, genAnd,   genOr, genXor, genAddrOf): fixed several bugs concerning xstack   (genModOneByte): fixed warning in MSVC * src/mcs51/main.c (): added comments * src/mcs51/peeph.def: changed 129 to 129.a, added 129.b, 129.c & 129.d
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@3735 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCopt.c
@@ -561,7 +561,7 @@ convertToFcall (eBBlock ** ebbs, int count)
           if (ic->op == '%' && isOperandLiteral(IC_RIGHT(ic)) &&
               IS_UNSIGNED(operandType(IC_LEFT(ic))))
             {
-              unsigned litVal = abs(operandLitValue(IC_RIGHT(ic)));
+              unsigned litVal = abs((int)operandLitValue(IC_RIGHT(ic)));
 
               // See if literal value is a power of 2.
               while (litVal && !(litVal & 1))
@@ -941,11 +941,6 @@ killDeadCode (ebbIndex * ebbi)
 		  if (ic->op == ADDRESS_OF)
 		    volLeft = FALSE;
               
-		  //if (ic->op == CAST && isOperandVolatile (IC_LEFT (ic), FALSE))
-		  //  {
-		  //    volRight = IS_SYMOP (IC_RIGHT (ic));
-		  //  }
-
 		  if (ic->next && ic->seqPoint == ic->next->seqPoint
 		      && (ic->next->op == '+' || ic->next->op == '-'))
 		    {
@@ -1170,9 +1165,6 @@ eBBlock **
 eBBlockFromiCode (iCode * ic)
 {
   ebbIndex *ebbi = NULL;
-  //eBBlock **ebbs = NULL;
-  //int count = 0;
-  //int saveCount = 0;
   int change = 1;
   int lchange = 0;
   int kchange = 0;
@@ -1182,7 +1174,6 @@ eBBlockFromiCode (iCode * ic)
   if (!ic)
     return NULL;
 
-  //count = 0;
   eBBNum = 0;
 
   /* optimize the chain for labels & gotos
